refactor: move interfaces to @types folder
