+++ web/src/components/PageHeader/index.tsx
@@ -1,33 +1,31 @@
-/* eslint-disable react/prop-types */
-/* eslint-disable react/destructuring-assignment */
 import React from 'react';
 import { Link } from 'react-router-dom';
+import { PageHeaderProps } from '../../@types';
 
 import logoImg from '../../assets/images/logo.svg';
 import backIcon from '../../assets/images/icons/back.svg';
 
 import './styles.css';
 
-interface PageHeaderProps {
-  title: string;
-  description?: string;
-  topBarTitle?: string;
-}
-
-const PageHeader: React.FC<PageHeaderProps> = (props) => {
+const PageHeader: React.FC<PageHeaderProps> = ({
+  title,
+  children,
+  description,
+  topBarTitle,
+}) => {
   return (
     <header className="page-header">
       <div className="top-bar-container">
         <Link to="/">
           <img src={backIcon} alt="Voltar" />
         </Link>
-        {props.topBarTitle && <h1 className="title">{props.topBarTitle}</h1>}
+        {topBarTitle && <h1 className="title">{topBarTitle}</h1>}
         <img src={logoImg} alt="Logo Proffy" />
       </div>
       <div className="header-content">
-        <strong>{props.title}</strong>
-        {props.description && <p>{props.description}</p>}
-        {props.children}
+        <strong>{title}</strong>
+        {description && <p>{description}</p>}
+        {children}
       </div>
     </header>
   );
